Update CCD size accoring to email from Eric Miller
On 11/16/17 Eric Miller writes:
Another day, another Arcus CCD size.  The width of the active imaging area is back to 2048 pixels (still 24 microns per pixel).  There are in fact 10 additional columns of "buffer" pixels, 5 columns on each side, but they are counted in the 0.824 mm of dead area between the last active column and the physical edge of the silicon.  These buffer columns are not usable because there are no corresponding columns in the framestore to transfer them to.

So the width of the imaging area is 2048x0.024 = 49.152 mm.  The width of the physical silicon is that plus 2x0.824 = 50.8 mm.  The smallest "gap" between active area of adjacent chips is 1.648 mm if they butt right up against each other.  In practice that's probably not an easy thing to do, nor a good idea, and the LL folks suggest an extra gap of 0.5 mm is conservatively good, for a total "gap" between active areas of 2.148 mm.  Mark probably knows better how close the CCDs can be safely and easily placed.

I would also use 1024 pixels for the height of the CCD imaging area instead of the 1039 I gave you recently.  Sorry for the confusion.