Update cross-dispersion FWHM to larger number
Randall Smith write in an email from 11/16/17:
However, a 10 arcsec FWHM is equivalent to 1sigma = 10 arcsec/2.355 = 4.25 arcsec, and the HEW is 1.35 sigma = 5.73 arcsec.  We never put a formal requirement on the cross-dispersion width, but the numbers that cosine has demonstrated are ~10 arcsec HEW, equivalent to 17.4 arcsec FWHM.  I expect them to improve on that in the next year (Athena needs it, and they’re funded to do that work), but we shouldn’t count on that.  My apologies for the confusion on this — it’s obvious I didn’t mention that the 10 arcsec was HEW, not FWHM, at some point.