Move boom up
email from Ed Hertz 6/11/2020:
The OD of the Boom is 1850 mm.  The batten lengths (in the XY plane where Z is the optical axis) are 1308 mm.  The length of each bay (in the Z direction) is 1350.5 mm.  There are a total of 8 bays and there are no battens at the top (closest to the Optical Channel) where the boom connects to the Forward Assembly.  The top-most bay begins 11.4 m from the focal plane (System 0,0,0 coordinate) or 600mm from the apex of the SPO sphere.  This should leave enough room for the rays from the outer row of SPO's to remain inside the boom entirely.